move argument parsing, PVC/DV error handling, and storageSpec helpers
